#1 group memberList and kicked
