+++ frontend/lib/main.dart
@@ -9,8 +9,11 @@ import 'package:frontend/screens/home/shorts/shorts_screen.dart';
 import 'package:frontend/screens/home/group/group_screen.dart';
 import 'package:frontend/screens/home/bookreport/bookreport_screen.dart';
 import 'package:frontend/screens/home/mypage/mypage_screen.dart';
-import 'package:frontend/screens/home/group/group_info_screen.dart';
-import 'package:frontend/screens/home/group/group_make_screen.dart';
+import 'package:frontend/screens/home/group/in_group/group_info_screen.dart';
+import 'package:frontend/screens/home/group/make_group/group_make_screen.dart';
+import 'package:frontend/screens/home/group/in_group/post/homework_list_screen.dart';
+import 'package:frontend/screens/home/group/in_group/post/notice_list_screen.dart';
+import 'package:frontend/screens/home/group/in_group/post/post_list_screen.dart';
 
 void main() async {
   runApp(const App());
@@ -75,6 +78,21 @@ final GoRouter router = GoRouter(
       name: 'group_make',
       path: '/group_make',
       builder: (context, state) => const GroupMakeScreen(),
+    ),
+    GoRoute(
+      name: 'homework_list',
+      path: '/homework_list',
+      builder: (context, state) => const HomeworkListScreen(),
+    ),
+    GoRoute(
+      name: 'notice_list',
+      path: '/notice_list',
+      builder: (context, state) => const NoticeListScreen(),
+    ),
+    GoRoute(
+      name: 'post_list',
+      path: '/post_list',
+      builder: (context, state) => const PostListScreen(),
     ),
   ],
 );
