+++ frontend/lib/screens/home/group/in_group/post/homework_list_screen.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:go_router/go_router.dart';
+
+//과제 목록 페이지
+
+class HomeworkListScreen extends StatefulWidget {
+  const HomeworkListScreen({super.key});
+
+  @override
+  State<HomeworkListScreen> createState() => _HomeworkListScreenState();
+}
+
+class _HomeworkListScreenState extends State<HomeworkListScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
