+++ frontend/lib/screens/home/group/in_group/post/notice_list_screen.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:go_router/go_router.dart';
+
+//공지사항 목록 페이지
+
+class NoticeListScreen extends StatefulWidget {
+  const NoticeListScreen({super.key});
+
+  @override
+  State<NoticeListScreen> createState() => _NoticeListScreenState();
+}
+
+class _NoticeListScreenState extends State<NoticeListScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
